+++ test/jtreg/generator/nestedInsideAnon/TestNestedInsideAnon.java
@@ -29,7 +29,7 @@ import java.lang.foreign.GroupLayout;
 import static org.testng.Assert.*;
 
 /*
- * @test id
+ * @test
  * @library /lib
  * @run main/othervm JtregJextract -t test.jextract.nested.anon nestedInsideAnon.h
  * @build TestNestedInsideAnon
