+++ test/jtreg/generator/nestedTypes/TestNestedTypesNames.java
@@ -31,7 +31,7 @@ import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
 /*
- * @test id
+ * @test
  * @library /lib
  * @run main/othervm JtregJextract -t test.jextract.nestedtypes.names nested_types_names.h
  * @build TestNestedTypesNames
