+++ test/lib/testlib/TestUtils.java
@@ -70,8 +70,6 @@ public class TestUtils {
             System.err.println("compiling jextracted sources @ " + sourcePath.toAbsolutePath());
             List<String> commands = new ArrayList<>();
             commands.add("-parameters");
-            commands.add("--source=22");
-            commands.add("--enable-preview");
             commands.add("-d");
             commands.add(outputDir.toAbsolutePath().toString());
             commands.addAll(files);
